+++ src/main/java/umc/study/Application.java
@@ -1,4 +1,4 @@
-package umc.spring;
+package umc.study;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
