+++ src/test/java/umc/study/ApplicationTests.java
@@ -1,4 +1,4 @@
-package umc.spring;
+package umc.study;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
